Split DB reader status hotspots (#4110)
