+++ changelog.d/unreleased/4110.internal.md
@@ -1,0 +1,18 @@
+---
+category: internal
+issues:
+  - 4110
+affected:
+  - src/CodeIndex/Database/DbReader.cs
+  - src/CodeIndex/Database/DbReader.FilesStatus.cs
+  - src/CodeIndex/Database/DbReader.SearchSymbolMatchJoins.cs
+  - src/CodeIndex/Database/DbReader.Status.cs
+---
+
+## English
+
+- **Database reader hotspots were split into narrower partial files (#4110)** — status aggregation and search-symbol ranking SQL now live in focused `DbReader` partials, reducing the base reader and file/status surfaces without changing query behavior.
+
+## 日本語
+
+- **database reader hotspot をより狭い partial file に分割しました (#4110)** — status aggregation と search-symbol ranking SQL を専用の `DbReader` partial に移し、query 挙動を変えずに base reader と file/status surface を小さくしました。
